Refactored Sudoku structure and unit tests.
A Sudoku is now simply a sudoku, that is, a matrix of 9 x 9. Eliminated the
attribute `initialValues` as its more related to the game than to the sudoku.
Added some new attributes like `rowVals`, `colVals`, `blockVals` to keep track
of the values contained in the sudoku in constant time, and also added
`emptyEntries` to check how many entries are missing in the sudoku.
Additional changes:

- Added a new method `NewSudoku` which returns a sudoku with some given values.

- Method `SetValue` now checks if the value isn't already repeated in the row,
  column, block it will be placed. If its repeated, it returns an error.

- Methods `GetRow`, `GetColumn`, `GetBlock` were removed as they're not
  relevant to the methods available to the sudoku structure.

- Methods `IsValidRow`, `IsValidColumn`, `IsValidBlock` were removed, this
  verifications are now performed when adding entries to the sudoku.

- Method `RemoveValue` was added, but its missing unitary testing.

- Method `IsComplety` simply checks that the amount of empty entries is zero.